f-footer@0.1.0 move base component files to its folder + fix test description
